Initial commit.
Login is implemented. Token is verified as working when making a request to get a non existing group.
There are tests that check that response for correct and incorrect token is handled.
In case of correct token path $HOME/.screenly is checked to have the token stored upon successful login.