Fixes `edge-app delete` (#235)
* fix: Handle missing edge app in get_app

The `get_app` function panicked when trying to access the first element
of the returned app list if the list was empty (e.g., app not found).

This commit adds a check for an empty list in `get_app` and returns
a `ResourceNotFound` error in that case. It also adds the
`ResourceNotFound` variant to the `CommandError` enum.

* Bumps deps

* run clippy

* Rolls back Cargo.lock

* Backward compatability

* Fixes compatability issues

* Rename ResourceNotFound to AppNotFound